Backend/requirements.txt
djangorestframework==3.16.0
django==5.2.4
django-cors-headers==4.6.0
netmiko==4.6.0
paramiko==3.5.0
textfsm==1.1.3
ntc-templates==8.1.0
transformers==4.44.2
# torch is optional; install only if running local models
# On Windows, install appropriate torch build manually if needed
peft==0.13.2
numpy
uvicorn
gunicorn
python-dotenv
requests

# PostgreSQL support
psycopg2-binary==2.9.9
dj-database-url==2.2.0
